pyproject.toml
[project]
name = "inflationagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.12",
    "gradio>=5.0.0",
    "lancedb>=0.25.3",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "openai>=2.11.0",
    "pylance>=1.0.0",
    "pysr>=0.16.0",
    "python-dotenv>=1.0.0",
    "scipy>=1.10.0",
    "sqlalchemy>=2.0.45",
    "tantivy>=0.25.1",
    "tiktoken>=0.5.0",
]
